cleanup; move IsCronJobPending, remove extraneous params
