+++ backend/controller/cronjobs/cronjobs.go
@@ -173,13 +173,10 @@ func (s *Service) scheduleCronJob(ctx context.Context, tx *Tx, job model.CronJob
 
 	logger.Tracef("Scheduling cron job %q async_call execution at %s", job.Key, nextAttemptForJob)
 	id, err := tx.db.CreateAsyncCall(ctx, cronsql.CreateAsyncCallParams{
-		ScheduledAt:       nextAttemptForJob,
-		Verb:              schema.RefKey{Module: job.Verb.Module, Name: job.Verb.Name},
-		Origin:            fmt.Sprintf("cron:%s", job.Key),
-		Request:           []byte(`{}`),
-		RemainingAttempts: 0,
-		Backoff:           0,
-		MaxBackoff:        0,
+		ScheduledAt: nextAttemptForJob,
+		Verb:        schema.RefKey{Module: job.Verb.Module, Name: job.Verb.Name},
+		Origin:      fmt.Sprintf("cron:%s", job.Key),
+		Request:     []byte(`{}`),
 	})
 	if err != nil {
 		return fmt.Errorf("failed to create async call for job %q: %w", job.Key, err)
