+++ backend/controller/cronjobs/sql/queries.sql
@@ -39,3 +39,13 @@ UPDATE cron_jobs
     last_execution = sqlc.arg('last_execution')::TIMESTAMPTZ,
     next_execution = sqlc.arg('next_execution')::TIMESTAMPTZ
   WHERE key = sqlc.arg('key')::cron_job_key;
+
+-- name: IsCronJobPending :one
+SELECT EXISTS (
+    SELECT 1
+    FROM cron_jobs j
+      INNER JOIN async_calls ac on j.last_async_call_id = ac.id
+    WHERE j.key = sqlc.arg('key')::cron_job_key
+      AND ac.scheduled_at > sqlc.arg('start_time')::TIMESTAMPTZ
+      AND ac.state = 'pending'
+) AS pending;
