+++ backend/controller/cronjobs/sql/queries.sql.go
@@ -149,6 +149,24 @@ func (q *Queries) GetUnscheduledCronJobs(ctx context.Context, startTime time.Tim
 	return items, nil
 }
 
+const isCronJobPending = `-- name: IsCronJobPending :one
+SELECT EXISTS (
+    SELECT 1
+    FROM cron_jobs j
+      INNER JOIN async_calls ac on j.last_async_call_id = ac.id
+    WHERE j.key = $1::cron_job_key
+      AND ac.scheduled_at > $2::TIMESTAMPTZ
+      AND ac.state = 'pending'
+) AS pending
+`
+
+func (q *Queries) IsCronJobPending(ctx context.Context, key model.CronJobKey, startTime time.Time) (bool, error) {
+	row := q.db.QueryRowContext(ctx, isCronJobPending, key, startTime)
+	var pending bool
+	err := row.Scan(&pending)
+	return pending, err
+}
+
 const updateCronJobExecution = `-- name: UpdateCronJobExecution :exec
 UPDATE cron_jobs
   SET last_async_call_id = $1::BIGINT,
