+++ backend/controller/sql/async_queries.sql
@@ -24,13 +24,3 @@ VALUES (
   @trace_context::jsonb
 )
 RETURNING id;
-
--- name: IsCronJobPending :one
-SELECT EXISTS (
-    SELECT 1
-    FROM cron_jobs j
-      INNER JOIN async_calls ac on j.last_async_call_id = ac.id
-    WHERE j.key = sqlc.arg('key')::cron_job_key
-      AND ac.scheduled_at > sqlc.arg('start_time')::TIMESTAMPTZ
-      AND ac.state = 'pending'
-) AS pending;
