+++ backend/controller/sql/queries.sql.go
@@ -2127,6 +2127,24 @@ func (q *Queries) InsertTimelineLogEvent(ctx context.Context, arg InsertTimeline
 	return err
 }
 
+const isCronJobPending = `-- name: IsCronJobPending :one
+SELECT EXISTS (
+    SELECT 1
+    FROM cron_jobs j
+      INNER JOIN async_calls ac on j.last_async_call_id = ac.id
+    WHERE j.key = $1::cron_job_key
+      AND ac.scheduled_at > $2::TIMESTAMPTZ
+      AND ac.state = 'pending'
+) AS pending
+`
+
+func (q *Queries) IsCronJobPending(ctx context.Context, key model.CronJobKey, startTime time.Time) (bool, error) {
+	row := q.db.QueryRowContext(ctx, isCronJobPending, key, startTime)
+	var pending bool
+	err := row.Scan(&pending)
+	return pending, err
+}
+
 const killStaleControllers = `-- name: KillStaleControllers :one
 WITH matches AS (
     UPDATE controller
